be/int: fix toNative for i8/i16/u8/u16

+++ src/dev/flang/be/interpreter/u8Value.java
@@ -122,7 +122,7 @@ public class u8Value extends Value
   @Override
   protected Object toNative()
   {
-    return this.u8Value();
+    return (byte)this.u8Value();
   }
 
 }
